Make RawKeyboard assert when no keys are in keysPressed when sending a key down event. (#63426)
This adds an assert in RawKeyboard that catches the case where it tries to send a key down event, but (after synchronizing modifiers) there are no keys in keysPressed. This state can occur if the modifier flags are not set properly for the platform.

Also prevents shortcuts attempting to handle a key down when no keys are pressed at the moment (which was causing a crash in release mode).