[flutter_tools] measure driver success and failure (#69226)
Currently we only have a test that measures if flutter drive can fail. Unfortunately it has been failing for multiple years based on the fact that the error is due to a missing Gradle config.

Add a smoke success test case, and update both tests to use the test API and the tester-device.