Add missed package version bump override (#82593)
In https://github.com/flutter/flutter/pull/82577, I landed a version
bump of package shelf and package flutter_template_images. This local
override should have been part of that patch but was missed in the
checkin.

The version bumps themselves happened in the previous patch; this just
ensures we don't accidentally 'upgrade' back to version 1.0.2.