Fixed leak and removed no-shuffle tag in widgets/app_overrides_test.dart (#88469)
This PR fixed the problem that has prevented app_overrides_test.dart being shuffled. Part of #85160.

This test is testing WidgetsApp.showPerformanceOverlayOverride and WidgetsApp.debugAllowBannerOverride.

The problem: There was a strong dependecy to the default ordering of the tests. The value of showPerformanceOverlayOverride or debugAllowBannerOverride was expected to be the value that was set in previous test.

The fix: Make the values be set to default values after each test:
WidgetsApp.showPerformanceOverlayOverride=false
WidgetsApp.debugAllowBannerOverride=true
Set the expected value at the start of the test, if the default value is not expected.