[tool] Fix BuildInfo.packagesPath doc comment (#103785)
Until recently, BuildInfo.packagesPath defaulted to `.packages` however,
the .packages file has been deprecated [1] and is being removed. In
https://github.com/flutter/flutter/pull/99677 the default was changed to
`.dart_tool/package_config.json` but the doc comment was accidentally
updated to `.dart_tool/packages`. This corrects it to the true default
value.

1: https://github.com/dart-lang/sdk/issues/48272

Spotted in the midst of some related cleanup relating to
https://github.com/flutter/flutter/issues/103775