Move RenderObjectElement.updateChildren to Element (#128458)
This method lived on RenderObjectElement because traditionally, it would be the only one that had to deal with multiple children. The method itself has nothing RenderObjectElement specific, though, and can also be used by any other Element subclass that has multiple children. We are introducing one of those in the near future to handle multiple top-level views.

This is a straight up copy&paste move, no changes have been applied to the code itself.